+++ App.navigation.tsx
@@ -2,9 +2,12 @@ import { NavigationContainer } from '@react-navigation/native'
 import { useTheme as reactNativePaperUseTheme } from 'react-native-paper'
 import { createMaterialBottomTabNavigator } from '@react-navigation/material-bottom-tabs'
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons'
-import { ThemeProvider, useTheme } from 'styled-components/native'
-import { ThemeModel } from '@core/styled/models'
+import { ThemeProvider } from 'styled-components/native'
+import { ActivityIndicator, ColorSchemeName, useColorScheme } from 'react-native'
+import { useSelector } from 'react-redux'
 import { theme as themeModel } from '@core/styled/theme'
+import { useRehydrateStore } from '@core/hooks/use-rehydrate-store.hook'
+import { STORAGE_KEY } from '@core/storage/storage.constants'
 import { Dashboard } from './src/dashboard/dashboard'
 import { COLOR } from './src/core/colors/colors.constants'
 import { SCREEN } from './src/navigation/navigation.constants'
@@ -12,25 +15,37 @@ import { Dictionary } from './src/dictionary/dictionary'
 import { genericShadow } from './src/core/shadow/shadow.constants'
 import { Playground } from './src/playground/playground'
 import { MoreNavigation } from './src/more/more.navigation'
+import { setDarkThemeEnabledAction } from './src/settings/store/settings.slice'
+import { darkThemeEnabledSelector } from './src/settings/store/settings.selectors'
 
 const Tab = createMaterialBottomTabNavigator()
 
 export const AppNavigation = () => {
   const reactNativePaperTheme = reactNativePaperUseTheme()
-  const theme = useTheme() as ThemeModel
   reactNativePaperTheme.colors.secondaryContainer = "transparent"
 
-  // const colorScheme: ColorSchemeName = useColorScheme()
-  const colorScheme = 'dark'
-  const backgroundColor = colorScheme === 'light' ? COLOR.WHITE : COLOR.DARK_SLATE_GREY
+  const { isPending } = useRehydrateStore(STORAGE_KEY.DARK_THEME_ENABLED, setDarkThemeEnabledAction)
+  const darkTheme = useSelector(darkThemeEnabledSelector)
 
-  return (
-    <ThemeProvider theme={themeModel[colorScheme ?? 'light']}>
+  const colorScheme: ColorSchemeName = useColorScheme()
+
+  const getBackgroundColor = () => darkTheme === -1 && colorScheme === 'light' || !darkTheme
+    ? COLOR.WHITE
+    : COLOR.DARK_SLATE_GREY
+
+  const getThemeToProvide = () => darkTheme === -1
+    ? themeModel[colorScheme ?? 'light']
+    : themeModel[darkTheme ? 'dark' : 'light']
+
+  return isPending ? (
+    <ActivityIndicator size="large" />
+  ) : (
+    <ThemeProvider theme={getThemeToProvide()}>
       <NavigationContainer>
         <Tab.Navigator
           inactiveColor={COLOR.SLATE_GREY}
           activeColor={COLOR.DODGER_BLUE}
-          barStyle={{ backgroundColor, ...genericShadow }}
+          barStyle={{ backgroundColor: getBackgroundColor(), ...genericShadow }}
           labeled={false}
           sceneAnimationType="shifting"
           sceneAnimationEnabled
@@ -64,16 +79,6 @@ export const AppNavigation = () => {
               ),
             }}
           />
-
-          {/* <Tab.Screen
-            name={SCREEN.MANIA}
-            component={Mania}
-            options={{
-              tabBarIcon: ({ color }) => (
-                <MaterialCommunityIcons name="web" color={color} size={26} />
-              ),
-            }}
-          /> */}
 
           <Tab.Screen
             name={SCREEN.MORE}